detector/finetune.py
import os
import sys
import logging
import argparse
import yaml
import json
from pathlib import Path
from typing import Dict, Any, List, Tuple, Optional
import time
from datetime import datetime

from ultralytics import YOLO

logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.StreamHandler(sys.stdout)
    ]
)
logger = logging.getLogger('finetune')

def parse_args() -> argparse.Namespace:
    parser = argparse.ArgumentParser(description='Finetune YOLOv8 waste detector on real imagery')
    parser.add_argument('--config', type=str, default='config/training_config.yaml',
                        help='Path to training configuration file')
    parser.add_argument('--weights', type=str, required=True,
                        help='Path to synthetic pre-trained weights')
    parser.add_argument('--real-data', type=str, required=True,
                        help='Path to real drone imagery dataset')
    parser.add_argument('--epochs', type=int, default=50,
                        help='Number of finetuning epochs')
    parser.add_argument('--batch-size', type=int, default=16,
                        help='Batch size')
    parser.add_argument('--learning-rate', type=float, default=0.001,
                        help='Initial learning rate')
    parser.add_argument('--device', type=str, default=None,
                        help='Training device (cpu, 0, 0,1,2,3, etc.)')
    parser.add_argument('--output', type=str, default='weights/finetuned.pt',
                        help='Output path for finetuned weights')
    return parser.parse_args()

def load_config(config_path: str) -> Dict[str, Any]:
    logger.info(f"Loading configuration from {config_path}")
    try:
        with open(config_path, 'r') as f:
            config = yaml.safe_load(f)
        return config
    except Exception as e:
        logger.error(f"Error loading configuration: {e}")
        sys.exit(1)

def setup_output_dirs(output_path: str) -> Path:
    output_dir = Path(output_path).parent
    output_dir.mkdir(parents=True, exist_ok=True)
    return output_dir

def finetune_model(
    weights_path: str,
    real_data_path: str,
    config: Dict[str, Any],
    args: argparse.Namespace
) -> Path:

    if not Path(weights_path).exists():
        logger.error(f"Pre-trained weights not found at {weights_path}")
        sys.exit(1)

    real_data_dir = Path(real_data_path)
    if not real_data_dir.exists():
        logger.error(f"Real dataset not found at {real_data_dir}")
        sys.exit(1)

    # TODO:

    logger.info(f"Loading pre-trained model from {weights_path}")
    model = YOLO(weights_path)

    image_size = config['hyperparameters']['image_size']

    device = args.device if args.device else 'cuda:0'

    finetune_args = {
        'data': real_data_path,
        'epochs': args.epochs,
        'batch': args.batch_size,
        'imgsz': image_size,
        'device': device,
        'lr0': args.learning_rate,
        'lrf': args.learning_rate / 10,
        'momentum': config['hyperparameters'].get('momentum', 0.937),
        'weight_decay': config['hyperparameters'].get('weight_decay', 0.0005),
        'warmup_epochs': 3.0,
        'warmup_momentum': 0.8,
        'warmup_bias_lr': 0.1,
        'box': config['hyperparameters'].get('box', 7.5),
        'cls': config['hyperparameters'].get('cls', 0.5),
        'project': 'istanbul-waste',
        'name': 'finetuned',
        'exist_ok': True,
        'save': True
    }

    start_time = time.time()

    logger.info(f"Starting finetuning for {args.epochs} epochs")
    logger.info(f"Learning rate: {args.learning_rate}")
    logger.info(f"Batch size: {args.batch_size}")
    logger.info(f"Device: {device}")

    # TODO: Uncomment and implement finetuning when real dataset is available

    logger.info("This is a placeholder script for future implementation.")
    logger.info("Real finetuning would be implemented when real drone imagery is available.")

    return Path(args.output)

def save_finetuning_results(
    output_path: Path,
    synthetic_weights: str,
    real_data_path: str,
    args: argparse.Namespace
) -> None:
    results = {
        'base_model': synthetic_weights,
        'real_dataset': real_data_path,
        'finetuned_model': str(output_path),
        'finetuning_parameters': {
            'epochs': args.epochs,
            'batch_size': args.batch_size,
            'learning_rate': args.learning_rate,
            'device': args.device
        },
        'timestamp': datetime.now().strftime('%Y-%m-%d %H:%M:%S')
    }

    results_path = output_path.parent / "finetuning_results.json"
    with open(results_path, 'w') as f:
        json.dump(results, f, indent=2)

    logger.info(f"Finetuning results saved to {results_path}")

def main() -> None:
    args = parse_args()

    config = load_config(args.config)

    log_level = getattr(logging, config['logging']['level'].upper())
    logger.setLevel(log_level)

    log_dir = Path('logs')
    log_dir.mkdir(exist_ok=True)
    file_handler = logging.FileHandler(log_dir / 'finetune.log', mode='w')
    file_handler.setFormatter(logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s'))
    logger.addHandler(file_handler)

    output_dir = setup_output_dirs(args.output)

    logger.info("Placeholder for future finetuning implementation")
    logger.info("This script would be implemented when real drone imagery becomes available")

if __name__ == "__main__":
    main()
