requirements.txt
airsim==1.9.0
numpy>=1.22.0
opencv-python>=4.7.0
pillow>=9.4.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
pyyaml>=6.0
tqdm>=4.65.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pycocotools>=2.0.6
jupyter>=1.0.0
notebook>=6.5.0
scikit-learn>=1.2.0
